delete & update pages fixed
compares values with database

+++ app/posts/delete.php
@@ -8,24 +8,28 @@ require __DIR__.'/../autoload.php';
 
 try {
 	$postId = $_GET['id'];
-	$username = $_SESSION['username'];
+	$username = getUserinfo($pdo);
+	$username = $username['username'];
 
-	$sql = 'DELETE FROM posts WHERE user_id = :username AND id = :postId';
-	$stmt = $pdo->prepare($sql);
-	$stmt->bindParam(':postId', $postId, PDO::PARAM_INT);
-	$stmt->bindParam(':username', $username, PDO::PARAM_STR);
-	$stmt->execute();
+	// Compare session username with db username
+	if ($username === $_SESSION['username']) {
 
-	if (!$stmt) {
-		$_SESSION['error'] = 'There was an error, please try again.';
-		redirect('/app/users/profile.php');
-		exit;
-	} else {
-		$_SESSION['success'] = 'Post was successfully deleted.';
-		redirect('/app/users/profile.php');
-		exit;
+			$sql = 'DELETE FROM posts WHERE user_id = :username AND id = :postId';
+			$stmt = $pdo->prepare($sql);
+			$stmt->bindParam(':postId', $postId, PDO::PARAM_INT);
+			$stmt->bindParam(':username', $username, PDO::PARAM_STR);
+			$stmt->execute();
+
+			if (!$stmt) {
+				$_SESSION['error'] = 'There was an error, please try again.';
+				redirect('/app/users/profile.php');
+				exit;
+			} else {
+				$_SESSION['success'] = 'Post was successfully deleted.';
+				redirect('/app/users/profile.php');
+				exit;
+		}
 	}
-
 } catch (Exception $e) {
 	echo 'Something went wrong with the connection: ' . $e->getMessage();
 }
